#7 #313 第七次实验 设置屏幕强制为竖屏
