+++ students/soft1706081301317/app/src/main/java/com/example/soft1706081301317sweepacticity/MainActivity.java
@@ -31,7 +31,7 @@ public class MainActivity extends AppCompatActivity {
       textView_2.setOnClickListener(new View.OnClickListener() {
           @Override
           public void onClick(View v) {
-              Intent  intent = new Intent(MainActivity.this,FunnyActivity.class);
+              Intent  intent = new Intent(MainActivity.this,HelpActivity.class);
               startActivity(intent);
           }
       });
