+++ students/soft1706081301317/app/src/main/java/com/example/soft1706081301317sweepacticity/WriteActivity.java
@@ -1,0 +1,38 @@
+package com.example.soft1706081301317sweepacticity;
+
+import android.app.Activity;
+import android.content.pm.ActivityInfo;
+import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+
+public class WriteActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_write);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        //设置强制竖屏
+        setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT);
+
+
+        Button button = findViewById(R.id.eat);
+        button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                EditText editText = findViewById(R.id.edit_title);
+                EditText editText1 = findViewById(R.id.edit_body);
+                editText.setText("");
+                editText1.setText("负能量没啦!睡个觉会更好点.晚安💗");
+            }
+        });
+    }
+
+}
