--- students/soft1706081301317/app/src/main/java/com/example/soft1706081301317sweepactivity/FunnyActivity.java
@@ -1,25 +1,0 @@
-package com.example.soft1706081301317sweepacticity;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.widget.Button;
-import android.widget.TextView;
-
-public class FunnyActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_funny);
-        TextView textView = findViewById(R.id.funny);
-        textView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent  intent = new Intent(FunnyActivity.this,HelpActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-}
