--- students/soft1706081301317/app/src/main/java/com/example/soft1706081301317sweepactivity/HelpActivity.java
@@ -1,25 +1,0 @@
-package com.example.soft1706081301317sweepacticity;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.widget.TextView;
-
-public class HelpActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_help);
-        TextView textView = (TextView)findViewById(R.id.back);
-        textView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(HelpActivity.this,MainActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-}
