--- students/soft1706081301317/app/src/main/java/com/example/soft1706081301317sweepactivity/SweepActivity.java
@@ -1,33 +1,0 @@
-package com.example.soft1706081301317sweepacticity;
-
-import android.content.Intent;
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-
-public class SweepActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sweep_acticity);
-        Button button = (Button)findViewById(R.id.save);
-        final EditText editText = findViewById(R.id.username);
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                SharedPreferences.Editor editor = getSharedPreferences("data",MODE_PRIVATE).edit();
-                editor.putString("username",editText.getText().toString());
-                editor.apply();
-                Intent intent = new Intent(SweepActivity.this,WeclomeActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-
-}
