--- students/soft1706081301317/app/src/main/res/layout/activity_funny.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:background="@drawable/p2"
-    tools:context=".FunnyActivity"
-    android:orientation="vertical">
-
-<Button
-    android:id="@+id/funny"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:text="我知道了"
-    android:textColor="#ffffff"
-    android:background="@android:color/transparent"
-    android:textSize="25dp"
-    />
-
-</LinearLayout>
