+++ students/soft1706081301317/app/src/main/res/layout/activity_help.xml
@@ -23,13 +23,13 @@
         android:layout_weight="0.7">
 
         <TextView
-            android:id="@+id/begin"
+
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="bottom"
             android:gravity="center"
-            android:text="左键单击展开"
-            android:textColor="#3399CC	"
+            android:text="最近有很多心事吧.."
+            android:textColor="#FFFFFF	"
             android:textSize="60dp"
             android:textStyle="bold" />
 
@@ -42,15 +42,35 @@
         android:layout_weight="0.9">
 
         <TextView
-            android:id="@+id/help"
+
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_gravity="top"
             android:gravity="center"
-            android:text="双击标记地雷"
-            android:textColor="#3399CC	"
+            android:text="写下你的负能量"
+            android:textColor="#FFFFFF	"
             android:textSize="60dp"
             android:textStyle="bold" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_weight="0.9">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="top"
+            android:gravity="center"
+            android:text="只有树洞知道"
+            android:textColor="#FFFFFF	"
+            android:textSize="60dp"
+            android:textStyle="bold" />
+
+
     </LinearLayout>
 
 </LinearLayout>
